fix: revert using unprivileged user
It's not that is impossible to run the daemon as an unprivileged user,
but for that to happen the device passed to the container must have the
container's UID or GID so it can access it. Because that requires some
coordination with the host, let's simply keep executing it as root.